+++ spec/grape/validations/params_scope_spec.rb
@@ -165,6 +165,19 @@ describe Grape::Validations::ParamsScope do
 
       expect(last_response.status).to eq(200)
     end
+
+    it do
+      subject.params do
+        requires :foo, as: :baz, type: Hash do
+          requires :bar, as: :qux
+        end
+      end
+      subject.get('/nested-renaming') { declared(params).to_json }
+      get '/nested-renaming', foo: { bar: 'any' }
+
+      expect(last_response.status).to eq(200)
+      expect(last_response.body).to eq('{"baz":{"qux":"any"}}')
+    end
   end
 
   context 'array without coerce type explicitly given' do